+++ docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Unreleased
+------------------------
+Fix #437: Hide the profile calendar if the module is not available for users
+
 1.5.7 (January 12, 2024)
 ------------------------
 - Fix #433: Fix memory usage on integrity check
@@ -14,7 +18,7 @@ Changelog
 -------------------------
 - Enh #424: Tests for `next` version
 - Enh #423: Add Create Button
-- Fix #419: Don't use loader for wall entry menu on open modal window 
+- Fix #419: Don't use loader for wall entry menu on open modal window
 - Enh #379: Allow managing of global event types from container side
 - Enh #382: Allow inviting by users who can manage a calendar entry