+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ------------------------
 Fix #437: Hide the profile calendar if the module is not available for users
+Fix #439: Attach ICS files to "forced invites"
 
 1.5.7 (January 12, 2024)
 ------------------------