+++ docs/CHANGELOG.md
@@ -5,7 +5,8 @@ Unreleased
 ------------------------
 Fix #437: Hide the profile calendar if the module is not available for users
 Fix #439: Attach ICS files to "forced invites"
-Fix #441: Send event update notifcation to invitees
+Fix #441: Send event update notification to invitees
+Fix #434: Fix adding of all space members with status
 
 1.5.7 (January 12, 2024)
 ------------------------